+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -183,6 +183,32 @@ public final class SentryClient implements ISentryClient {
     return event;
   }
 
+  @Override
+  public void captureUserFeedback(UserFeedback userFeedback) {
+    Objects.requireNonNull(userFeedback, "SentryEvent is required.");
+
+    options.getLogger().log(SentryLevel.DEBUG, "Capturing userFeedback: %s", userFeedback.getEventId());
+
+    try {
+      final SentryEnvelope envelope = buildEnvelope(userFeedback);
+      connection.send(envelope);
+    } catch (IOException e) {
+      options.getLogger().log(SentryLevel.WARNING, e, "Capturing user feedback %s failed.", userFeedback.getEventId());
+    }
+  }
+
+  private SentryEnvelope buildEnvelope(@NotNull UserFeedback userFeedback) {
+    final List<SentryEnvelopeItem> envelopeItems = new ArrayList<>();
+
+    final SentryEnvelopeItem userFeedbackItem = SentryEnvelopeItem.fromUserFeedback(options.getSerializer(), userFeedback);
+    envelopeItems.add(userFeedbackItem);
+
+    final SentryEnvelopeHeader envelopeHeader =
+      new SentryEnvelopeHeader(userFeedback.getEventId(), options.getSdkVersion());
+
+    return new SentryEnvelope(envelopeHeader, envelopeItems);
+  }
+
   /**
    * Updates the session data based on the event, hint and scope data
    *