+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.1.2
+
+* feat: Manually capturing User Feedback
+* Enhancement: Set environment to "production" by default.
+* Enhancement: Make public the Breadcrumb constructor that accepts a Date #1012
+
 # 3.1.1
 
 * fix: Prevent Logback and Log4j2 integrations from re-initializing Sentry when Sentry is already initialized